@@ -1,23 +1,25 @@
 module github.com/sunerpy/requests
 
-go 1.24.3
+go 1.25.5
 
 require (
-	github.com/agiledragon/gomonkey/v2 v2.13.0
-	github.com/bytedance/sonic v1.13.3
-	github.com/stretchr/testify v1.10.0
+	github.com/agiledragon/gomonkey/v2 v2.14.0
+	github.com/bytedance/sonic v1.14.2
+	github.com/leanovate/gopter v0.2.11
+	github.com/stretchr/testify v1.11.1
 )
 
 require (
-	github.com/bytedance/sonic/loader v0.2.4 // indirect
-	github.com/cloudwego/base64x v0.1.5 // indirect
+	github.com/bytedance/gopkg v0.1.3 // indirect
+	github.com/bytedance/sonic/loader v0.4.0 // indirect
+	github.com/cloudwego/base64x v0.1.6 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/klauspost/cpuid/v2 v2.2.10 // indirect
+	github.com/klauspost/cpuid/v2 v2.3.0 // indirect
 	github.com/kr/text v0.2.0 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/twitchyliquid64/golang-asm v0.15.1 // indirect
-	golang.org/x/arch v0.18.0 // indirect
-	golang.org/x/sys v0.33.0 // indirect
+	golang.org/x/arch v0.23.0 // indirect
+	golang.org/x/sys v0.39.0 // indirect
 	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
