@@ -1,3 +1,4 @@
+// Package requests provides a simple and easy-to-use HTTP client library for Go.
 package requests
 
 import (
@@ -8,16 +9,18 @@ import (
 	"fmt"
 	"net"
 	"net/http"
+	"net/http/cookiejar"
 	"net/url"
 	"sync"
 	"time"
 
-	"github.com/sunerpy/requests/models"
+	"github.com/sunerpy/requests/internal/client"
+	"github.com/sunerpy/requests/internal/models"
 )
 
 const (
 	idleConnTimeout         = 90 * time.Second
-	dnsResloveTimeout       = 15 * time.Second
+	dnsResolveTimeout       = 15 * time.Second
 	defaultDisableKeepAlive = false
 	defaultMaxIdleConns     = 100
 )
@@ -25,15 +28,16 @@ const (
 var (
 	http1TransportPool  sync.Pool
 	http2TransportPool  sync.Pool
+	sessionPool         sync.Pool
 	defaultHTTP2Enabled = false
 	defaultHTTP2Lock    sync.Mutex
 	defaultSess         Session
 )
 
-// 初始化 Transport 池并创建默认 Session
+// Initialize Transport pools and create default Session
 func init() {
 	http1TransportPool = sync.Pool{
-		New: func() interface{} {
+		New: func() any {
 			return &http.Transport{
 				MaxIdleConns:        defaultMaxIdleConns,
 				MaxIdleConnsPerHost: defaultMaxIdleConns,
@@ -45,20 +49,30 @@ func init() {
 		},
 	}
 	http2TransportPool = sync.Pool{
-		New: func() interface{} {
+		New: func() any {
 			return &http.Transport{
 				MaxIdleConns:        defaultMaxIdleConns,
 				MaxIdleConnsPerHost: defaultMaxIdleConns,
 				IdleConnTimeout:     idleConnTimeout,
 				DisableKeepAlives:   defaultDisableKeepAlive,
 				ForceAttemptHTTP2:   true,
-				// TLSNextProto 保持默认，支持 HTTP/2
+			}
+		},
+	}
+	sessionPool = sync.Pool{
+		New: func() any {
+			return &defaultSession{
+				headers:      make(http.Header, 4),
+				keepAlive:    true,
+				maxIdleConns: defaultMaxIdleConns,
+				idleTimeout:  idleConnTimeout,
 			}
 		},
 	}
 	defaultSess = NewSession()
 }
 
+// GetTransport returns a transport from the pool.
 func GetTransport(enableHTTP2 bool) *http.Transport {
 	if enableHTTP2 {
 		return http2TransportPool.Get().(*http.Transport)
@@ -66,6 +80,7 @@ func GetTransport(enableHTTP2 bool) *http.Transport {
 	return http1TransportPool.Get().(*http.Transport)
 }
 
+// PutTransport returns a transport to the pool.
 func PutTransport(transport *http.Transport) {
 	if transport == nil {
 		return
@@ -77,7 +92,58 @@ func PutTransport(transport *http.Transport) {
 	}
 }
 
-// SetHTTP2Enabled 设置全局 HTTP/2 启用状态，并更新默认 Session
+// AcquireSession gets a Session from the pool.
+// Remember to call ReleaseSession when done.
+func AcquireSession() Session {
+	s := sessionPool.Get().(*defaultSession)
+	defaultHTTP2Lock.Lock()
+	useHTTP2 := defaultHTTP2Enabled
+	defaultHTTP2Lock.Unlock()
+	s.useHTTP2 = useHTTP2
+	if useHTTP2 {
+		s.client = &http.Client{Transport: http2TransportPool.Get().(*http.Transport)}
+	} else {
+		s.client = &http.Client{Transport: http1TransportPool.Get().(*http.Transport)}
+	}
+	return s
+}
+
+// ReleaseSession returns a Session to the pool.
+func ReleaseSession(sess Session) {
+	if sess == nil {
+		return
+	}
+	s, ok := sess.(*defaultSession)
+	if !ok {
+		return
+	}
+	// Return transport to pool
+	if tr, ok := s.client.Transport.(*http.Transport); ok {
+		if tr.ForceAttemptHTTP2 {
+			http2TransportPool.Put(tr)
+		} else {
+			http1TransportPool.Put(tr)
+		}
+	}
+	// Reset session state
+	s.baseURL = ""
+	s.timeout = 0
+	s.idleTimeout = idleConnTimeout
+	s.proxyURL = nil
+	s.dnsServers = nil
+	s.authHeader = ""
+	s.bearerToken = ""
+	s.keepAlive = true
+	s.maxIdleConns = defaultMaxIdleConns
+	// Clear headers but keep the map
+	for k := range s.headers {
+		delete(s.headers, k)
+	}
+	s.client = nil
+	sessionPool.Put(s)
+}
+
+// SetHTTP2Enabled sets the global HTTP/2 enabled state.
 func SetHTTP2Enabled(enabled bool) {
 	defaultHTTP2Lock.Lock()
 	defer defaultHTTP2Lock.Unlock()
@@ -85,17 +151,20 @@ func SetHTTP2Enabled(enabled bool) {
 	defaultSess = defaultSess.WithHTTP2(enabled)
 }
 
-// IsHTTP2Enabled 获取全局 HTTP/2 启用状态
+// IsHTTP2Enabled returns the global HTTP/2 enabled state.
 func IsHTTP2Enabled() bool {
 	defaultHTTP2Lock.Lock()
 	defer defaultHTTP2Lock.Unlock()
 	return defaultHTTP2Enabled
 }
 
+// Client is the HTTP client interface.
 type Client interface {
 	Do(req *Request) (*models.Response, error)
 	Clone() Client
 }
+
+// Session extends Client with session management capabilities.
 type Session interface {
 	Client
 	WithBaseURL(base string) Session
@@ -103,16 +172,23 @@ type Session interface {
 	WithProxy(proxyURL string) Session
 	WithDNS(dnsServers []string) Session
 	WithHeader(key, value string) Session
+	WithHeaders(headers map[string]string) Session
 	WithBasicAuth(username, password string) Session
+	WithBearerToken(token string) Session
 	WithHTTP2(enabled bool) Session
 	WithKeepAlive(enabled bool) Session
 	WithMaxIdleConns(maxIdle int) Session
+	WithIdleTimeout(d time.Duration) Session
+	WithCookieJar(jar http.CookieJar) Session
 	Close() error
 	Clear() Session
 }
+
+// defaultSession implements the Session interface.
 type defaultSession struct {
 	baseURL      string
 	timeout      time.Duration
+	idleTimeout  time.Duration
 	proxyURL     *url.URL
 	headers      http.Header
 	client       *http.Client
@@ -122,9 +198,10 @@ type defaultSession struct {
 	clientLock   sync.Mutex
 	dnsServers   []string
 	authHeader   string
+	bearerToken  string
 }
 
-// NewSession 创建一个新的 Session，使用对应的 Transport 池
+// NewSession creates a new Session with default settings.
 func NewSession() Session {
 	defaultHTTP2Lock.Lock()
 	defer defaultHTTP2Lock.Unlock()
@@ -134,12 +211,14 @@ func NewSession() Session {
 	} else {
 		transport = http1TransportPool.Get().(*http.Transport)
 	}
+	jar, _ := cookiejar.New(nil)
 	return &defaultSession{
 		headers:      http.Header{},
-		client:       &http.Client{Transport: transport},
+		client:       &http.Client{Transport: transport, Jar: jar},
 		useHTTP2:     defaultHTTP2Enabled,
 		keepAlive:    true,
 		maxIdleConns: defaultMaxIdleConns,
+		idleTimeout:  idleConnTimeout,
 	}
 }
 
@@ -150,15 +229,22 @@ func (s *defaultSession) WithBaseURL(base string) Session {
 	return s
 }
 
-// This function `WithTimeout` in the `defaultSession` struct is setting the timeout duration for the
-// session. It acquires a lock on the client, sets the timeout duration to the provided value `d`, and
-// the timeout duration is a request timeout.
 func (s *defaultSession) WithTimeout(d time.Duration) Session {
 	s.clientLock.Lock()
 	defer s.clientLock.Unlock()
 	s.timeout = d
 	if s.client != nil {
 		s.client.Timeout = d
+	}
+	return s
+}
+
+func (s *defaultSession) WithIdleTimeout(d time.Duration) Session {
+	s.clientLock.Lock()
+	defer s.clientLock.Unlock()
+	s.idleTimeout = d
+	if tr, ok := s.client.Transport.(*http.Transport); ok {
+		tr.IdleConnTimeout = d
 	}
 	return s
 }
@@ -184,7 +270,6 @@ func (s *defaultSession) WithProxy(proxyURL string) Session {
 func (s *defaultSession) WithDNS(dnsServers []string) Session {
 	s.clientLock.Lock()
 	defer s.clientLock.Unlock()
-	// 设置自定义的 DNS 服务器
 	s.dnsServers = dnsServers
 	if tr, ok := s.client.Transport.(*http.Transport); ok {
 		tr.DialContext = func(ctx context.Context, network, address string) (net.Conn, error) {
@@ -192,7 +277,7 @@ func (s *defaultSession) WithDNS(dnsServers []string) Session {
 				return customDial(ctx, network, address, s.dnsServers)
 			}
 			dialer := &net.Dialer{
-				Timeout:   dnsResloveTimeout,
+				Timeout:   dnsResolveTimeout,
 				DualStack: true,
 			}
 			return dialer.DialContext(ctx, network, address)
@@ -201,13 +286,11 @@ func (s *defaultSession) WithDNS(dnsServers []string) Session {
 	return s
 }
 
-// customDial 使用自定义 DNS 服务器解析并连接目标地址
 func customDial(ctx context.Context, network, address string, dnsServers []string) (net.Conn, error) {
 	dialer := &net.Dialer{
-		Timeout:   dnsResloveTimeout,
+		Timeout:   dnsResolveTimeout,
 		DualStack: true,
 	}
-	// 创建自定义的 Resolver
 	resolver := &net.Resolver{
 		PreferGo: true,
 		Dial: func(ctx context.Context, network, address string) (net.Conn, error) {
@@ -247,6 +330,14 @@ func (s *defaultSession) WithBasicAuth(username, password string) Session {
 	return s
 }
 
+func (s *defaultSession) WithBearerToken(token string) Session {
+	s.clientLock.Lock()
+	defer s.clientLock.Unlock()
+	s.bearerToken = token
+	s.headers.Set("Authorization", fmt.Sprintf("Bearer %s", token))
+	return s
+}
+
 func (s *defaultSession) WithHeader(key, value string) Session {
 	s.clientLock.Lock()
 	defer s.clientLock.Unlock()
@@ -254,7 +345,15 @@ func (s *defaultSession) WithHeader(key, value string) Session {
 	return s
 }
 
-// WithHTTP2 设置当前 session 是否使用 HTTP/2
+func (s *defaultSession) WithHeaders(headers map[string]string) Session {
+	s.clientLock.Lock()
+	defer s.clientLock.Unlock()
+	for k, v := range headers {
+		s.headers.Set(k, v)
+	}
+	return s
+}
+
 func (s *defaultSession) WithHTTP2(enabled bool) Session {
 	s.clientLock.Lock()
 	defer s.clientLock.Unlock()
@@ -263,13 +362,11 @@ func (s *defaultSession) WithHTTP2(enabled bool) Session {
 	}
 	s.useHTTP2 = enabled
 	if tr, ok := s.client.Transport.(*http.Transport); ok {
-		// 将当前 Transport 放回相应的池中
 		if tr.ForceAttemptHTTP2 {
 			http2TransportPool.Put(tr)
 		} else {
 			http1TransportPool.Put(tr)
 		}
-		// 获取新的 Transport
 		if enabled {
 			tr = http2TransportPool.Get().(*http.Transport)
 		} else {
@@ -311,29 +408,39 @@ func (s *defaultSession) WithMaxIdleConns(maxIdle int) Session {
 	return s
 }
 
+func (s *defaultSession) WithCookieJar(jar http.CookieJar) Session {
+	s.clientLock.Lock()
+	defer s.clientLock.Unlock()
+	s.client.Jar = jar
+	return s
+}
+
 func (s *defaultSession) Clone() Client {
 	s.clientLock.Lock()
 	defer s.clientLock.Unlock()
-	// 获取新的 Transport
 	var transport *http.Transport
 	if s.useHTTP2 {
 		transport = http2TransportPool.Get().(*http.Transport)
 	} else {
 		transport = http1TransportPool.Get().(*http.Transport)
 	}
+	jar, _ := cookiejar.New(nil)
 	newClient := &http.Client{
 		Transport: transport,
+		Jar:       jar,
 	}
 	newHeaders := s.headers.Clone()
 	newSession := &defaultSession{
 		baseURL:      s.baseURL,
 		timeout:      s.timeout,
+		idleTimeout:  s.idleTimeout,
 		proxyURL:     s.proxyURL,
 		headers:      newHeaders,
 		client:       newClient,
 		useHTTP2:     s.useHTTP2,
 		keepAlive:    s.keepAlive,
 		maxIdleConns: s.maxIdleConns,
+		bearerToken:  s.bearerToken,
 	}
 	if !s.useHTTP2 {
 		newSession.client.Transport.(*http.Transport).TLSNextProto = make(map[string]func(string, *tls.Conn) http.RoundTripper)
@@ -355,7 +462,6 @@ func (s *defaultSession) Close() error {
 	return nil
 }
 
-// Clear 清空当前 Session 的所有配置并恢复到默认状态
 func (s *defaultSession) Clear() Session {
 	s.clientLock.Lock()
 	defer s.clientLock.Unlock()
@@ -363,12 +469,39 @@ func (s *defaultSession) Clear() Session {
 	s.useHTTP2 = false
 	s.keepAlive = true
 	s.maxIdleConns = defaultMaxIdleConns
-	s.timeout = idleConnTimeout
+	s.timeout = 0
+	s.idleTimeout = idleConnTimeout
 	s.dnsServers = nil
+	s.bearerToken = ""
+	s.authHeader = ""
+	s.headers = http.Header{}
+	jar, _ := cookiejar.New(nil)
 	s.client = &http.Client{
-		Transport: GetTransport(false), // 重置为默认 Transport
+		Transport: GetTransport(false),
+		Jar:       jar,
 	}
 	return s
+}
+
+// resolveURL resolves the request URL against the session's base URL.
+func (s *defaultSession) resolveURL(reqURL *url.URL) (*url.URL, error) {
+	if s.baseURL == "" {
+		return reqURL, nil
+	}
+	base, err := url.Parse(s.baseURL)
+	if err != nil {
+		return nil, err
+	}
+	return base.ResolveReference(reqURL), nil
+}
+
+// applyHeaders copies headers from source to destination.
+func applyHeaders(dst, src http.Header) {
+	for k, vals := range src {
+		for _, v := range vals {
+			dst.Add(k, v)
+		}
+	}
 }
 
 func (s *defaultSession) Do(req *Request) (*models.Response, error) {
@@ -378,13 +511,9 @@ func (s *defaultSession) Do(req *Request) (*models.Response, error) {
 	if req.Context == nil {
 		req.Context = context.Background()
 	}
-	finalURL := req.URL
-	if s.baseURL != "" {
-		base, err := url.Parse(s.baseURL)
-		if err != nil {
-			return nil, err
-		}
-		finalURL = base.ResolveReference(req.URL)
+	finalURL, err := s.resolveURL(req.URL)
+	if err != nil {
+		return nil, err
 	}
 	ctx := req.Context
 	if s.timeout > 0 {
@@ -392,23 +521,42 @@ func (s *defaultSession) Do(req *Request) (*models.Response, error) {
 		ctx, cancel = context.WithTimeout(ctx, s.timeout)
 		defer cancel()
 	}
-	httpReq, err := http.NewRequestWithContext(ctx, req.Method, finalURL.String(), req.Body)
+	httpReq, err := http.NewRequestWithContext(ctx, req.Method.String(), finalURL.String(), req.Body)
 	if err != nil {
 		return nil, err
 	}
-	for k, vals := range s.headers {
-		for _, v := range vals {
-			httpReq.Header.Add(k, v)
-		}
+	// Fast path: skip header iteration if no session headers
+	if len(s.headers) > 0 {
+		applyHeaders(httpReq.Header, s.headers)
 	}
-	for k, vals := range req.Headers {
-		for _, v := range vals {
-			httpReq.Header.Add(k, v)
-		}
+	// Fast path: skip header iteration if no request headers
+	if len(req.Headers) > 0 {
+		applyHeaders(httpReq.Header, req.Headers)
 	}
-	resp, err := s.client.Do(httpReq)
+	resp, err := s.client.Do(httpReq) //nolint:bodyclose // body is closed in models.NewResponse
 	if err != nil {
 		return nil, err
 	}
 	return models.NewResponse(resp, finalURL.String())
 }
+
+// DoFast is an optimized version of Do for simple requests without baseURL or timeout.
+// Use this when you need maximum performance and don't need session-level configuration.
+func (s *defaultSession) DoFast(req *Request) (*models.Response, error) {
+	httpReq, err := http.NewRequestWithContext(req.Context, req.Method.String(), req.URL.String(), req.Body)
+	if err != nil {
+		return nil, err
+	}
+	if len(req.Headers) > 0 {
+		httpReq.Header = req.Headers
+	}
+	resp, err := s.client.Do(httpReq) //nolint:bodyclose // body is closed in models.NewResponse
+	if err != nil {
+		return nil, err
+	}
+	return models.NewResponse(resp, req.URL.String())
+}
+
+// Request is an alias for client.Request - the unified request type.
+// Use NewRequestBuilder or convenience constructors (NewGet, NewPost, etc.) to create requests.
+type Request = client.Request
