@@ -17,11 +17,116 @@ func BuildURL(base string, params *Values) (string, error) {
 	return u.String(), nil
 }
 
-func ParseParams(data interface{}) *Values {
+// FastBuildURL builds a URL with query parameters using FastValues for better performance.
+// This is optimized for single-threaded scenarios.
+func FastBuildURL(base string, params *FastValues) (string, error) {
+	u, err := url.Parse(base)
+	if err != nil {
+		return "", err
+	}
+	if params != nil && params.Len() > 0 {
+		u.RawQuery = params.Encode()
+	}
+	return u.String(), nil
+}
+
+// BuildURLFast builds a URL with query parameters from a map for maximum performance.
+// Avoids intermediate allocations by building the query string directly.
+func BuildURLFast(base string, params map[string]string) (string, error) {
+	if len(params) == 0 {
+		return base, nil
+	}
+	u, err := url.Parse(base)
+	if err != nil {
+		return "", err
+	}
+	// Build query string directly
+	var sb strings.Builder
+	sb.Grow(len(params) * 32) // Estimate ~32 bytes per param
+	first := true
+	for k, v := range params {
+		if !first {
+			sb.WriteByte('&')
+		}
+		sb.WriteString(url.QueryEscape(k))
+		sb.WriteByte('=')
+		sb.WriteString(url.QueryEscape(v))
+		first = false
+	}
+	u.RawQuery = sb.String()
+	return u.String(), nil
+}
+
+// NewURLParams creates a new thread-safe Values for URL parameters.
+// This is the recommended default for most use cases as it provides
+// safe concurrent access using sync.RWMutex.
+//
+// For high-performance single-threaded scenarios, use NewURLParamsUnsafe instead.
+func NewURLParams() *Values {
+	return NewValues()
+}
+
+// NewURLParamsUnsafe creates a new non-thread-safe FastValues for URL parameters.
+// This provides better performance in single-threaded scenarios by avoiding
+// lock overhead.
+//
+// WARNING: This is NOT safe for concurrent access. Use NewURLParams for
+// thread-safe operations.
+func NewURLParamsUnsafe() *FastValues {
+	return NewFastValues()
+}
+
+// AcquireURLParams gets a thread-safe Values from the pool for URL parameters.
+// Remember to call ReleaseURLParams when done.
+// This combines object pooling with thread-safety for optimal performance
+// in concurrent scenarios.
+func AcquireURLParams() *Values {
+	return AcquireValues()
+}
+
+// ReleaseURLParams returns a thread-safe Values to the pool.
+func ReleaseURLParams(v *Values) {
+	ReleaseValues(v)
+}
+
+// AcquireURLParamsUnsafe gets a non-thread-safe FastValues from the pool.
+// Remember to call ReleaseURLParamsUnsafe when done.
+//
+// WARNING: This is NOT safe for concurrent access. Use AcquireURLParams for
+// thread-safe operations.
+func AcquireURLParamsUnsafe() *FastValues {
+	return AcquireFastValues()
+}
+
+// ReleaseURLParamsUnsafe returns a non-thread-safe FastValues to the pool.
+func ReleaseURLParamsUnsafe(v *FastValues) {
+	ReleaseFastValues(v)
+}
+
+// BuildURLWithQuery builds a URL with query parameters from url.Values.
+func BuildURLWithQuery(base string, query url.Values) (string, error) {
+	u, err := url.Parse(base)
+	if err != nil {
+		return "", err
+	}
+	if len(query) > 0 {
+		// Merge with existing query parameters
+		existingQuery := u.Query()
+		for k, vs := range query {
+			for _, v := range vs {
+				existingQuery.Add(k, v)
+			}
+		}
+		u.RawQuery = existingQuery.Encode()
+	}
+	return u.String(), nil
+}
+
+func ParseParams(data any) *Values {
 	return parseValues(data)
 }
 
-func parseValues(data interface{}) *Values {
+func parseValues(data any) *Values {
 	p := NewValues()
 	switch v := data.(type) {
 	case string:
@@ -38,7 +143,7 @@ func parseValues(data interface{}) *Values {
 		parseMapFloat(v, p)
 	case map[string][]float64:
 		parseMapFloatSlice(v, p)
-	case map[string]interface{}:
+	case map[string]any:
 		parseMapInterface(v, p)
 	}
 	return p
@@ -98,7 +203,7 @@ func parseMapFloatSlice(data map[string][]float64, p *Values) {
 	}
 }
 
-func parseMapInterface(data map[string]interface{}, p *Values) {
+func parseMapInterface(data map[string]any, p *Values) {
 	for key, value := range data {
 		switch v := value.(type) {
 		case string:
@@ -115,7 +220,7 @@ func parseMapInterface(data map[string]interface{}, p *Values) {
 			parseMapFloatSlice(map[string][]float64{key: v}, p)
 		case bool:
 			p.Add(key, strconv.FormatBool(v))
-		case []interface{}:
+		case []any:
 			for _, item := range v {
 				parseSingleInterface(key, item, p)
 			}
@@ -123,7 +228,7 @@ func parseMapInterface(data map[string]interface{}, p *Values) {
 	}
 }
 
-func parseSingleInterface(key string, value interface{}, p *Values) {
+func parseSingleInterface(key string, value any, p *Values) {
 	switch v := value.(type) {
 	case string:
 		p.Add(key, v)
